+++ packages/react/.storybook/preview.tsx
@@ -2,17 +2,21 @@ import { ArgTypes, Description, Title } from '@storybook/addon-docs';
 import type { Preview } from '@storybook/react';
 import { useEffect } from 'react';
 
-// constants
-import { BACKGROUND_COLOR_DARK, BACKGROUND_COLOR_LIGHT } from '@/constants';
+// styles
+import './styles/index.css';
+
+// theme
+import theme from '@/theme';
 
 const preview: Preview = {
   decorators: [
     (Story, context) => {
       useEffect(() => {
-        document.body.style.background =
-          context.globals.theme === 'light'
-            ? BACKGROUND_COLOR_LIGHT
-            : BACKGROUND_COLOR_DARK;
+        const background: string | null = theme.tokens.getByName(context.globals.theme === 'light' ? 'colors.backgroundLight.500' : 'colors.backgroundDark.500')?.value || null;
+
+        if (background) {
+          document.body.style.background = background;
+        }
       }, [context.globals.theme]);
       return <Story />;
     },