+++ packages/react/src/constants/Colors.ts
@@ -1,2 +1,0 @@
-export const BACKGROUND_COLOR_DARK = '#1a202c';
-export const BACKGROUND_COLOR_LIGHT = '#fff';