+++ packages/react/src/hooks/il8n/useIl8n.ts
@@ -1,0 +1,6 @@
+import { i18n as Il8n } from 'i18next';
+import { getI18n } from 'react-i18next';
+
+export default function useIl8n(): Il8n {
+  return getI18n();
+}