+++ packages/react/src/theme.ts
@@ -5,12 +5,69 @@ export default createSystem(
   defineConfig({
     globalCss: {
       html: {
+        bg: 'bg',
+        colorPalette: {
+          _dark: 'primaryDark',
+          _light: 'primaryLight',
+        },
         lineHeight: '1.5',
       },
     },
     theme: {
       semanticTokens: {
         colors: {
+          backgroundDark: {
+            bg: {
+              value: '{colors.backgroundDark.500}',
+            },
+            contrast: {
+              value: '{colors.whiteAlpha.800}', // text on solid bg
+            },
+            emphasized: {
+              value: '{colors.backgroundDark.700}', // hover/active
+            },
+            fg: {
+              value: '{colors.backgroundDark.500}', // text color
+            },
+            focusRing: {
+              value: '{colors.backgroundDark.500}', // focus ring
+            },
+            muted: {
+              value: '{colors.backgroundDark.500}', // muted bg
+            },
+            solid: {
+              value: '{colors.backgroundDark.500}', // main button bg
+            },
+            subtle: {
+              value: '{colors.backgroundDark.200}', // subtle bg
+            },
+          },
+          backgroundLight: {
+            bg: {
+              value: '{colors.backgroundLight.500}',
+            },
+            contrast: {
+              value: '{colors.gray.600}', // text on solid bg
+            },
+            emphasized: {
+              value: '{colors.backgroundLight.700}', // hover/active
+            },
+            fg: {
+              value: '{colors.backgroundLight.500}', // text color
+            },
+            focusRing: {
+              value: '{colors.backgroundLight.500}', // focus ring
+            },
+            muted: {
+              value: '{colors.backgroundLight.500}', // muted bg
+            },
+            solid: {
+              value: '{colors.backgroundLight.500}', // main button bg
+            },
+            subtle: {
+              value: '{colors.backgroundLight.200}', // subtle bg
+            },
+          },
           primaryDark: {
             contrast: {
               value: '{colors.gray.800}', // text on solid bg
@@ -31,7 +88,7 @@ export default createSystem(
               value: '{colors.primaryDark.500}', // main button bg
             },
             subtle: {
-              value: '{colors.primaryDark.200}', // subtle bg
+              value: '{colors.primaryDark.500}', // subtle bg
             },
           },
           primaryLight: {
@@ -42,36 +99,48 @@ export default createSystem(
               value: '{colors.primaryLight.700}',
             },
             fg: {
-              value: '{colors.primaryLight.100}',
+              value: '{colors.primaryLight.500}',
             },
             focusRing: {
               value: '{colors.primaryLight.500}',
             },
             muted: {
-              value: '{colors.primaryLight.100}',
+              value: '{colors.primaryLight.500}',
             },
             solid: {
               value: '{colors.primaryLight.500}',
             },
             subtle: {
-              value: '{colors.primaryLight.200}',
+              value: '{colors.primaryLight.500}',
             },
           },
         },
       },
       tokens: {
         colors: {
-          algorand: {
-            50: { value: '#000000' },
-            100: { value: '#000000' },
-            200: { value: '#000000' },
-            300: { value: '#000000' },
-            400: { value: '#1a1a1a' },
-            500: { value: '#333333' },
-            600: { value: '#4d4d4d' },
-            700: { value: '#666666' },
-            800: { value: '#ffffff' },
-            900: { value: '#ffffff' },
+          backgroundDark: {
+            50: { value: '#83878e' },
+            100: { value: '#6c7079' },
+            200: { value: '#565b65' },
+            300: { value: '#414651' },
+            400: { value: '#2d333e' },
+            500: { value: '#1a202c' },
+            600: { value: '#151a25' },
+            700: { value: '#10151e' },
+            800: { value: '#0c0f17' },
+            900: { value: '#070a11' },
+          },
+          backgroundLight: {
+            50: { value: '#fcfbfa' },
+            100: { value: '#fbfbfa' },
+            200: { value: '#fbfaf9' },
+            300: { value: '#faf9f8' },
+            400: { value: '#faf9f7' },
+            500: { value: '#f9f8f6' },
+            600: { value: '#d9d8d6' },
+            700: { value: '#b9b8b7' },
+            800: { value: '#9a9a98' },
+            900: { value: '#7d7c7b' },
           },
           primaryDark: {
             50: { value: '#F6E9FF' },
@@ -96,18 +165,6 @@ export default createSystem(
             700: { value: '#66026F' },
             800: { value: '#520159' },
             900: { value: '#3D0143' },
-          },
-          voi: {
-            50: { value: '#d9c7f7' },
-            100: { value: '#bb9af1' },
-            200: { value: '#ac84ee' },
-            300: { value: '#9d6deb' },
-            400: { value: '#8e57e8' },
-            500: { value: '#702ae2' },
-            600: { value: '#591abf' },
-            700: { value: '#4e17a9' },
-            800: { value: '#441492' },
-            900: { value: '#2f0e65' },
           },
         },
         fonts: {