+++ packages/react/src/types/index.ts
@@ -2,6 +2,8 @@ export * from './accounts';
 export * from './buttons';
 export * from './common';
 export * from './forms';
+export * from './il8n';
 export * from './layouts';
 export * from './loaders';
 export * from './strings';
+export * from './typography';