+++ packages/react/src/types/typography/THeadingProps.ts
@@ -1,0 +1,8 @@
+import type { HeadingProps } from '@chakra-ui/react';
+
+// types
+import type { IBaseComponentProps } from '@/types';
+
+type THeadingProps = IBaseComponentProps & HeadingProps;
+
+export default THeadingProps;