+++ packages/react/src/types/typography/TTextProps.ts
@@ -1,0 +1,8 @@
+import type { TextProps } from '@chakra-ui/react';
+
+// types
+import type { IBaseComponentProps } from '@/types';
+
+type TTextProps = IBaseComponentProps & TextProps;
+
+export default TTextProps;