+++ packages/react/CHANGELOG.md
@@ -1,3 +1,38 @@
+# [@kibisis/react-v1.4.0](https://github.com/kibis-is/toolkit/compare/@kibisis/react-v1.3.1...@kibisis/react-v1.4.0) (2025-07-04)
+
+
+### Features
+
+* **@kibisis/react:** expose il8n translate function and instance via… ([#11](https://github.com/kibis-is/toolkit/issues/11)) ([0434624](https://github.com/kibis-is/toolkit/commit/04346245062f4039f53c545da51363c6e5e21017))
+
+# [@kibisis/react-v1.3.1](https://github.com/kibis-is/toolkit/compare/@kibisis/react-v1.3.0...@kibisis/react-v1.3.1) (2025-07-03)
+
+
+### Bug Fixes
+
+* **@kibisis/react:** external translations not loading ([#10](https://github.com/kibis-is/toolkit/issues/10)) ([4e44b11](https://github.com/kibis-is/toolkit/commit/4e44b11fd048cffe010fc71edc0ebf8f7a346414))
+
+# [@kibisis/react-v1.3.0](https://github.com/kibis-is/toolkit/compare/@kibisis/react-v1.2.0...@kibisis/react-v1.3.0) (2025-07-03)
+
+
+### Features
+
+* **@kibisis/react:** add typography ([#9](https://github.com/kibis-is/toolkit/issues/9)) ([21aef96](https://github.com/kibis-is/toolkit/commit/21aef96512776ae1f7432cdfd15dba3d7ad922de))
+
+# [@kibisis/react-v1.2.0](https://github.com/kibis-is/toolkit/compare/@kibisis/react-v1.1.0...@kibisis/react-v1.2.0) (2025-07-03)
+
+
+### Features
+
+* **@kibisis/react:** expose default theme ([#8](https://github.com/kibis-is/toolkit/issues/8)) ([13dcafb](https://github.com/kibis-is/toolkit/commit/13dcafbda91cd38fee2ebc29d93f6c20b61216df))
+
+# [@kibisis/react-v1.1.0](https://github.com/kibis-is/toolkit/compare/@kibisis/react-v1.0.0...@kibisis/react-v1.1.0) (2025-07-03)
+
+
+### Features
+
+* **@kibisis/react:** add optional translations ([#7](https://github.com/kibis-is/toolkit/issues/7)) ([b4970bb](https://github.com/kibis-is/toolkit/commit/b4970bb1203cb532a4f9ac496e56ec169012c8f9))
+
 # @kibisis/react-v1.0.0 (2025-07-03)